backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "fastapi>=0.136.1",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "aiosqlite>=0.19.0",
]
